+++ Packs/FeedRecordedFuture/Integrations/FeedRecordedFuture/CHANGELOG.md
@@ -1,5 +1,5 @@
 ## [Unreleased]
-
+Renamed the **Sub-Feeds** parameter to **Services** in the instance configuration.
 
 ## [20.3.3] - 2020-03-18
 Added the **DomainGlob** indicator type to the feed's output, which might be applied when domains are returned.